+++ modules/overrides/interface.go
@@ -92,6 +92,7 @@ type Interface interface {
 	MaxMetricsDuration(userID string) time.Duration
 	DedicatedColumns(userID string) backend.DedicatedColumns
 	UnsafeQueryHints(userID string) bool
+	LeftPadTraceIDs(userID string) bool
 	CostAttributionMaxCardinality(userID string) uint64
 	CostAttributionDimensions(userID string) map[string]string